Set Content-Length explicitely for HEAD requests
If a HEAD request is being made with curl it will wait forever for some
content if neither a body nor a Content-Length header is received.
Setting the length header explicitely will curl terminate as expected.

This change also removes the CORS headers for HEAD requests as HEAD
isn't a whitelisted cors method anyway. It could also be added back and
whitelisted if that is wished.